creating a new class DisplayMenu
