+++ src/main/java/pl/console/project/App.java
@@ -13,7 +13,7 @@ public class App {
     public static void main(String[] args) {
 
         Menu.intro();
-        Menu.mainMenu(); //comment this line to check how the methods below work
+//        Menu.mainMenu(); //comment this line to check how the methods below work
 
 //        //Display users ListMeal
 //        STDOUT.info(Repository.getInstance().getListMeal().toString());
