+++ src/main/java/pl/console/project/menu/SearchMenu.java
@@ -2,8 +2,11 @@ package pl.console.project.menu;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import pl.console.project.model.ListMeal;
+import pl.console.project.model.Meal;
 
 import java.util.InputMismatchException;
+import java.util.List;
 import java.util.Scanner;
 
 class SearchMenu {
